Implement IE emulation for www.youtube.com by sending the IE 10 user-agent
The MetroUI Internet Explorer on Windows 8 does not offer Flash and so youtube
has to offer their videos as html5 for it.
At the moment it is sufficient to send the IE10 user-agent, but in the future
some javascript modifications may become necessary.

This is based on a review by Raylan Givens on mozilla addons.