Add PluginPermissionChanger module for managing per site plugin permissions
This is basically a thin wrapper around the firefox nsIPermissionManager service.
It provides simple access to set the status of the plugin and a callback mechanism
that notifies you of changed permissions. The therefore necessary observers are
managed automatically and removed on unload of the module.
You have to create an instance of the class per site and plugin.