Add function that ensures the right video quality for the viewport size is loaded
Because we now change the size of the container that contains the video before any
youtube js has the chance to load, it sometimes does not use the highes possible
video quality.

This new function waits until the video starts and then tells the youtube js which
is the highest possible quality setting for the chosen video size.

This uses the youtube API function setPlaybackQuality(suggestedQuality:String):Void