Add function to run code in page scope
This is necessary, because since firefox 33 access to the page's js scope
is very limited and does not allow for complex modifications.
We can bypass this by executing the code that makes the changes itself in the
page scope. This is done by appending a script tag with the serialized
code to the page dom.