Refactor video resolution enforcement to support ff33+; remove fexp code
The unsafeWindow API got restricted in firefox 33 and so some operations didn't
work anymore. Basically it is no longer allowed to expose objects from the
privileged addon scope to the page scope. All objects have to be cloned using
the structured clone algorithm, which does not support meta-programming like
properties. (similar to serializing to json)
This commit introdues an approach that works around this limitation by using
the function introduced in the last commit to run serialized functions in the
page scope, which then in turn are able to construct complex objects in the page
scope.

All static values, simple objects and functions are exported using the new
mechanisms (createObjectIn, cloneInto, exportFunction) provided since firefox 33,
only the code that constructs the youtube config injector is run in the page
scope using the function described above.

As a side effect, this commit removes the code that removed the youtube
experiment which was responsible for fixed video sizes, because it is now
implemented permanently and cannot be disabled anymore.
Another solution for the fixed video sizes will be introduced in the next commit.