Partially fix the problem with fixed size videos
Youtube permanently switched to using fixed sized videos. This results in the
video being shown too small when using the option to increase the size of the
video player.
This commit introduces some css that pertially fixes the problem: The video
itself is now displayed properly, however the progress bar and annotations
are not. It is based on an idea by Yonezpt in the issue linked below.

There is currently no simple solution to this problem, because all possible
solutions either break other functionality or involve injecting code into
the minified youtube player, which is very hard to maintain. (and ugly)

See here for more details: https://github.com/YePpHa/YouTubeCenter/issues/1083