Remove old API and iframe video loading methods
The Player-API method does no longer work, so it is removed.
The iframe method can still be used by clicking the button, but the method was
removed, because it breaks some youtube features and is no longer necessary
with the ie method.

Additionally now that most (all?) videos work by default, keeping old video
loading methods just does not seem appropriate, as this just confuses users.