Add new loading methods "Force HTML5-Test" and "Disable"
The former forces the usage of the html5 player by setting the propery f2 of the
yt PREF-cookie to 40000000, which is the same values set by youtube.com/html5.
This even works when flash is installed and enabled.
The second method just does nothing, which is appropriate for current firefox
versions without flash, because these will use the html5 player by default.