Fix issue that changed settings are only applied after restart
This broke in firefox 30, because we relied on unspecified, implementation-
defined behaviour: Changing the variable that is used when creating the
PageMod, see https://bugzilla.mozilla.org/show_bug.cgi?id=1027605 .

This is now fixed by recreating the PageMods after a setting is changed. This
does however not affect already opened tabs, which have to be reloaded for the
changes to take effect.