Fix progress bar and annotation display when resizing the player
This is a follow up commit of 8a0e787d61aa1cd61d625da97c45ccbea051a43b. It
fixes the display of the video controls and annotations that were not
resized correctly together with the player.

It works by deleting the browser internal function matchMedia(), which the
player uses to determine the size it should display its contents. If this
function does not exist, it falls back to using the size of the containing
div as a size reference. Because this also works for the video itself, the
css that was previously used to at least force the video to fill the player
is removed with this commit.

This fix was originally developed by Yonezpt for YouTubeCenter, thanks!
See here for details: https://github.com/YePpHa/YouTubeCenter/issues/1634