Hide resize options from menu when auto-resizing is disabled
Resizing only works correctly when the matchMedia function is deleted (see
previous commit), which is only done when the automatic resizing is enabled,
because we do not want to interfere with the player when the user is not
using the resize option. So we just remove the possibility the use manual
resizing when not using automatic resizing.