Improve responsiveness of the "start videos paused" option
This is done by using the playing DOM event to detect when the video has
started. It is then immediately paused using the video element without
waiting for the youtube api to finish loading (onYouTubePlayerReady).
This results in an immediate stop of the video without an audible split
second of playback.

This commit also removes the code that fixed the title by removing the play
icon when the video was paused, because youtube no longer uses this icon.